+++ src/test/kotlin/blackjack/testmachine/NormalCardMachine.kt
@@ -2,8 +2,8 @@ package blackjack.testmachine
 
 import blackjack.model.deck.Card
 import blackjack.model.deck.CardMachine
-import blackjack.model.deck.CardNumber
-import blackjack.model.deck.Pattern
+import blackjack.model.util.CardNumber
+import blackjack.model.util.Pattern
 
 class NormalCardMachine : CardMachine {
     override fun shuffle(cards: List<Card>): List<Card> {
